Klocwork_LS1021_ASF: Fixing klocwork issues.
This patch fixes the klocwork issues present in
ASF_SDK_LS1021A_V1-1.

CR: ENGR00331753
Reviewed-on: http://git.am.freescale.net:8181/29948
Tested-by: Sandeep Malik <Sandeep.Malik@freescale.com>

+++ asfmodule/asfffp/driver/asfreasm.c
@@ -909,7 +909,7 @@ static inline struct asf_reasmCb_s *asfIPv6ReasmFindOrCreateCb(
 #endif
 static inline int asfIPv4CheckFragInfo(struct sk_buff *skb,
 				       int *offset,  unsigned int *flags, unsigned int *ulSegLen,
-				       int *pIhl, unsigned int ulVSGId)
+				       unsigned short *pIhl, unsigned int ulVSGId)
 {
 	struct iphdr *iph;
 	int ihl;
@@ -968,7 +968,7 @@ static inline int asfIPv4CheckFragInfo(struct sk_buff *skb,
 
 static inline int asfIPv6CheckFragInfo(struct sk_buff *skb,
 				       int *offset,  unsigned int *flags, unsigned int *ulSegLen,
-				       int *pIhl, unsigned int ulVSGId)
+				       unsigned short *pIhl, unsigned int ulVSGId)
 {
 	struct ipv6hdr *ipv6h = ipv6_hdr(skb);
 	struct frag_hdr *fhdr = (struct frag_hdr *)skb_transport_header(skb);
@@ -1015,7 +1015,7 @@ static inline int asfIPv6CheckFragInfo(struct sk_buff *skb,
 
 static inline struct sk_buff  *asfFragHandle(struct asf_reasmCb_s *pCb,
 						 struct sk_buff *skb, unsigned int *pOffset,
-						 unsigned int flags, unsigned int *pLen, unsigned int ihl,
+						 unsigned int flags, unsigned int *pLen, unsigned short ihl,
 						 bool *bReasmDone, struct asf_fragInfo_s **frag, unsigned char *option) {
 	struct asf_fragInfo_s *pFrag, *pFragPrev;
 	unsigned int ulSegMap;
@@ -1522,7 +1522,7 @@ struct sk_buff  *asfIpv4Defrag(unsigned int ulVSGId,
 	unsigned int ulOffset = 0;
 	unsigned int flags = 0;
 	unsigned int ulSegLen = 0;
-	unsigned int ihl;
+	unsigned short ihl;
 	struct asf_fragInfo_s *pFrag;
 	struct sk_buff *pTempSkb;
 	struct sk_buff *pHeadSkb;
